refactor: move table_definitions under database
