+++ src/macaron/slsa_analyzer/database_store.py
@@ -8,16 +8,16 @@ from datetime import datetime
 from macaron import __version__
 from macaron.config.defaults import defaults
 from macaron.database.database_manager import DatabaseManager
-from macaron.output_reporter.results import Record
-from macaron.slsa_analyzer.analyze_context import AnalyzeContext
-from macaron.slsa_analyzer.checks.check_result import CheckResultType
-from macaron.slsa_analyzer.table_definitions import (
+from macaron.database.table_definitions import (
     AnalysisTable,
     CheckFactsTable,
     CheckResultTable,
     RepositoryAnalysis,
     SLSARequirement,
 )
+from macaron.output_reporter.results import Record
+from macaron.slsa_analyzer.analyze_context import AnalyzeContext
+from macaron.slsa_analyzer.checks.check_result import CheckResultType
 from macaron.util import logger
